create tooltip helper
