+++ frontend/fingertips-frontend/components/organisms/SparklineChart/SparklineChart.test.tsx
@@ -1,14 +1,6 @@
 import { render, screen } from '@testing-library/react';
-import {
-  SparklineChart,
-  sparklineTooltipContent,
-} from '@/components/organisms/SparklineChart/index';
+import { SparklineChart } from '@/components/organisms/SparklineChart/index';
 import { expect } from '@jest/globals';
-import { SparklineLabelEnum } from '@/components/organisms/BarChartEmbeddedTable';
-import {
-  BenchmarkComparisonMethod,
-  BenchmarkOutcome,
-} from '@/generated-sources/ft-api-client';
 
 describe('SparklineChart', () => {
   const mockValue = [48];
@@ -32,130 +24,5 @@ describe('SparklineChart', () => {
         'highcharts-react-component-barChartEmbeddedTable'
       )
     ).toBeInTheDocument();
-  });
-
-  describe('sparklineTooltipContent', () => {
-    it('should return the category "Benchmark" prefix and an empty benchmark label when benchmark data is present', () => {
-      const benchmarkOutcome = BenchmarkOutcome.Similar;
-      const benchmarkComparisonMethod =
-        BenchmarkComparisonMethod.CIOverlappingReferenceValue95;
-
-      const result = sparklineTooltipContent(
-        benchmarkOutcome,
-        SparklineLabelEnum.Benchmark,
-        benchmarkComparisonMethod
-      );
-
-      expect(result).toEqual({
-        benchmarkLabel: '',
-        category: 'Benchmark: ',
-        comparisonLabel: '',
-      });
-    });
-
-    it('should return the category "Group" prefix when group data is present', () => {
-      const benchmarkOutcome = BenchmarkOutcome.NotCompared;
-      const benchmarkComparisonMethod =
-        BenchmarkComparisonMethod.CIOverlappingReferenceValue95;
-
-      const result = sparklineTooltipContent(
-        benchmarkOutcome,
-        SparklineLabelEnum.Group,
-        benchmarkComparisonMethod
-      );
-
-      expect(result).toEqual({
-        benchmarkLabel: '',
-        category: 'Group: ',
-        comparisonLabel: '',
-      });
-    });
-
-    it('should return "Similar to England" benchmark label when the benchmark outcome is Similar', () => {
-      const benchmarkOutcome = BenchmarkOutcome.Similar;
-      const benchmarkComparisonMethod =
-        BenchmarkComparisonMethod.CIOverlappingReferenceValue95;
-
-      const result = sparklineTooltipContent(
-        benchmarkOutcome,
-        SparklineLabelEnum.Area,
-        benchmarkComparisonMethod
-      );
-
-      expect(result).toEqual({
-        benchmarkLabel: 'Similar to England',
-        category: '',
-        comparisonLabel: '(95.0%)',
-      });
-    });
-
-    it('should return comparison label of "95%" when benchmark comparison method is CIOverlappingReferenceValue95', () => {
-      const benchmarkOutcome = BenchmarkOutcome.Similar;
-      const benchmarkComparisonMethod =
-        BenchmarkComparisonMethod.CIOverlappingReferenceValue95;
-
-      const result = sparklineTooltipContent(
-        benchmarkOutcome,
-        SparklineLabelEnum.Area,
-        benchmarkComparisonMethod
-      );
-
-      expect(result).toEqual({
-        benchmarkLabel: 'Similar to England',
-        category: '',
-        comparisonLabel: '(95.0%)',
-      });
-    });
-
-    it('should not return a comparison label or benchmark label when the benchmark outcome method of "Not compared" is passed in', () => {
-      const benchmarkOutcome = BenchmarkOutcome.NotCompared;
-      const benchmarkComparisonMethod = BenchmarkComparisonMethod.Unknown;
-
-      const result = sparklineTooltipContent(
-        benchmarkOutcome,
-        SparklineLabelEnum.Benchmark,
-        benchmarkComparisonMethod
-      );
-
-      expect(result).toEqual({
-        benchmarkLabel: '',
-        category: 'Benchmark: ',
-        comparisonLabel: '',
-      });
-    });
-
-    it('should return tooltip for quintiles with no CI% shown', () => {
-      const benchmarkOutcome = BenchmarkOutcome.Worse;
-      const benchmarkComparisonMethod = BenchmarkComparisonMethod.Quintiles;
-
-      const result = sparklineTooltipContent(
-        benchmarkOutcome,
-        SparklineLabelEnum.Area,
-        benchmarkComparisonMethod
-      );
-
-      expect(result).toEqual({
-        benchmarkLabel: 'Worse quintile',
-        category: '',
-        comparisonLabel: '',
-      });
-    });
-
-    it('should return Not compared when the benchmark outcome method of "Not compared" is passed in', () => {
-      const benchmarkOutcome = BenchmarkOutcome.NotCompared;
-      const benchmarkComparisonMethod = BenchmarkComparisonMethod.Unknown;
-
-      const result = sparklineTooltipContent(
-        benchmarkOutcome,
-        SparklineLabelEnum.Area,
-        benchmarkComparisonMethod
-      );
-
-      expect(result).toEqual({
-        benchmarkLabel: 'Not compared',
-        category: '',
-        comparisonLabel: '',
-      });
-    });
   });
 });
